generic embedTemplate
