+++ lib/engines.js
@@ -1,5 +1,6 @@
 'use strict';
 var _ = require('lodash'),
+  template = require('./template'),
   engines = {},
   supported = ['nunjucks', 'jade'];
 
@@ -15,7 +16,7 @@ _.map(supported, function(name) {
 });
 
 // add embed filters and mixins
-engines.nunjucks.addFilter('embed', require('./nunjucks').embedFilter);
+engines.nunjucks.addFilter('embed', template.embed);
 
 module.exports = {
   engines: engines,
@@ -26,7 +27,12 @@ module.exports = {
   },
 
   jade: {
-    render: engines.jade.renderFile.bind(engines.jade)
+    render: function (file, options) {
+      options = options || {};
+      // add embed to the locals
+      options.embed = template.embed;
+      return engines.jade.renderFile(file, options);
+    }
   }
 };
