+++ lib/template.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var _ = require('lodash'),
-  embed = require('../embed');
+  embed = require('./embed');
 
 /**
  * Tries to embed another template into this template
@@ -9,7 +9,7 @@ var _ = require('lodash'),
  * @param {string} name
  * @param {{}} [extraData]
  */
-function embedFilter(data, name, extraData) {
+function embedTemplate(data, name, extraData) {
   if (!_.isString(name)) {
     return '';
   }
@@ -25,4 +25,4 @@ function embedFilter(data, name, extraData) {
   return embed.render(name, data);
 }
 
-exports.embedFilter = embedFilter;
+exports.embed = embedTemplate;
